Added new stuff (#5)
* Updated some stuff?

* Added getNewID func and constructButton func

getNewID func - passes filename and it will generate an id key (B001) and the index to insert the record at in the file. (in a list)

constructButton func - pass to it a 2D list containing the button name and button label. It will auto print to you the buttons according to the specified format given.
It receives column amt, 2d buttonList, moveleftamt from the centre aligned and resolution of the page

* Added u_popup

* Added insert header and insert line function

* Added some stuff